+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.1]
 ### Added
 - [Supervisor] Authentication fail reasons are stored in request attributes
-
+- [Core] User can now be injected directly to controller
 ### Fixed
 - [Supervisor] User is now added to RequestContext in stateless supervisors